requirements.txt
pandas
numpy
requests
pytz
python-dotenv
ta
scikit-learn
psycopg[binary]